Working on improving the RSL evaluation
- implemented a new RSL method where the JOINs to the securities and
  exchanges tables only happen in the outer query instead of in the
  inner
- also moved all the RSL evaluation functionality out the the
  QuoteDataService and into its own service class (there's enough code
  there that ´this is justified, and there's also litte to no chance
  that any of the RSL code might actually be re-used somewhere else)
- will probably do the same for the performance evakluation...
- with the data I have (25,000 to 30,000 price entries) I am not seeing
  a real difference in performance (calling this in ARC takes 150 to
  250 ms) with each of the implementation
- since I think the new implementation is cleaner it stays
