--- src/dtos/evaluation.dto.ts
@@ -1,19 +1,0 @@
-type RSLevyResponseData = {
-    securityISIN: string;
-    securityName: string;
-    instrumentType: string;
-    exchangeName: string;
-    newestWeeklyClose: Date;
-    rslValue: number;
-};
-
-type RSLevyWeeklyData = {
-    securityISIN: string;
-    securityName: string;
-    instrumentType: string;
-    exchangeName: string;
-    lastDayOfWeek: Date;
-    lastPriceOfWeek: number;
-};
-
-export { RSLevyResponseData, RSLevyWeeklyData };
