+++ src/dtos/index.ts
@@ -1,4 +1,3 @@
-export { RSLevyResponseData, RSLevyWeeklyData } from "./evaluation.dto";
 export { CreateExchangeRequest } from "./exchange.dto";
 export { AddQuoteDataRequest, NewestDatesOptions, NewestSharePriceDateDTO } from "./quote.dto";
 export { CreateSecurityRequest } from "./security.dto";
