+++ src/services/index.ts
@@ -1,4 +1,5 @@
 export { EvaluationService } from "./evaluation-service";
 export { ExchangeService } from "./exchange-service";
 export { QuoteDataService } from "./quote-service";
+export { RSLevyResponseData, RSLevyService } from "./rsleval.service";
 export { SecuritiesService } from "./security-service";
